Move state reachability analyser to its own class
